Split notifications into separate classes + move namespace
