+++ src/Umbraco.Core/Events/ICancelableNotification.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Umbraco.
+// See LICENSE for more details.
+
+namespace Umbraco.Cms.Core.Events
+{
+    public interface ICancelableNotification : INotification
+    {
+        bool Cancel { get; set; }
+    }
+}
