+++ src/Umbraco.Infrastructure/PropertyEditors/FileUploadPropertyEditor.cs
@@ -14,7 +14,7 @@ using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Serialization;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Strings;
-using Umbraco.Cms.Infrastructure.Services;
+using Umbraco.Cms.Infrastructure.Services.Notifications;
 using Umbraco.Extensions;
 
 namespace Umbraco.Cms.Core.PropertyEditors
@@ -125,7 +125,6 @@ namespace Umbraco.Cms.Core.PropertyEditors
                     yield return _mediaFileSystem.GetRelativePath(str2);
             }
         }
-
 
         public void Handle(CopiedNotification<IContent> notification)
         {
