+++ src/Umbraco.Infrastructure/Services/Notifications/CopiedNotification.cs
@@ -1,0 +1,22 @@
+// Copyright (c) Umbraco.
+// See LICENSE for more details.
+
+using Umbraco.Cms.Core.Events;
+
+namespace Umbraco.Cms.Infrastructure.Services.Notifications
+{
+    public class CopiedNotification<T> : ObjectNotification<T> where T : class
+    {
+        public CopiedNotification(T original, T copy, int parentId, EventMessages messages) : base(original, messages)
+        {
+            Copy = copy;
+            ParentId = parentId;
+        }
+
+        public T Original => Target;
+
+        public T Copy { get; }
+
+        public int ParentId { get; }
+    }
+}
