+++ src/Umbraco.Infrastructure/Services/Notifications/RolledBackNotification.cs
@@ -1,0 +1,16 @@
+// Copyright (c) Umbraco.
+// See LICENSE for more details.
+
+using Umbraco.Cms.Core.Events;
+
+namespace Umbraco.Cms.Infrastructure.Services.Notifications
+{
+    public class RolledBackNotification<T> : ObjectNotification<T> where T : class
+    {
+        public RolledBackNotification(T target, EventMessages messages) : base(target, messages)
+        {
+        }
+
+        public T Entity => Target;
+    }
+}
