+++ src/Umbraco.Infrastructure/Services/Notifications/SortedNotification.cs
@@ -1,0 +1,17 @@
+// Copyright (c) Umbraco.
+// See LICENSE for more details.
+
+using System.Collections.Generic;
+using Umbraco.Cms.Core.Events;
+
+namespace Umbraco.Cms.Infrastructure.Services.Notifications
+{
+    public class SortedNotification<T> : EnumerableObjectNotification<T>
+    {
+        public SortedNotification(IEnumerable<T> target, EventMessages messages) : base(target, messages)
+        {
+        }
+
+        public IEnumerable<T> SortedEntities => Target;
+    }
+}
