+++ src/Umbraco.Infrastructure/Services/Notifications/SortingNotification.cs
@@ -1,0 +1,17 @@
+// Copyright (c) Umbraco.
+// See LICENSE for more details.
+
+using System.Collections.Generic;
+using Umbraco.Cms.Core.Events;
+
+namespace Umbraco.Cms.Infrastructure.Services.Notifications
+{
+    public class SortingNotification<T> : CancelableEnumerableObjectNotification<T>
+    {
+        public SortingNotification(IEnumerable<T> target, EventMessages messages) : base(target, messages)
+        {
+        }
+
+        public IEnumerable<T> SortedEntities => Target;
+    }
+}
